Minor fixes for payment and birthday lists

+++ lib/utilities/format_date/full_date_time.dart
@@ -5,4 +5,5 @@ String fullDate(DateTime dateTime) => DateFormat("MMMM d'th' - hh:mm aa")?.forma
 String fullDateWithYear(DateTime dateTime) => DateFormat("EEEE, MMMM d'th' yyyy hh:mm aa")?.format(dateTime);
 String date(DateTime dateTime) => DateFormat("EEEE, MMMM d'th' yyyy")?.format(dateTime);
 String time(DateTime dateTime) => DateFormat("hh:mm aa")?.format(dateTime);
-String money(int money) => NumberFormat.simpleCurrency(name: 'USD', locale: "en_US", decimalDigits: 00).format(money);
+String money(double money) =>
+    NumberFormat.simpleCurrency(name: 'USD', locale: "en_US", decimalDigits: 00).format(money);
